Catch Up

+++ Forms/Gen3/ProfileManager3.cpp
@@ -1,6 +1,6 @@
 /*
  * This file is part of PokéFinder
- * Copyright (C) 2017 by Admiral_Fish, bumba, and EzPzStreamz
+ * Copyright (C) 2017-2019 by Admiral_Fish, bumba, and EzPzStreamz
  *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
@@ -17,11 +17,14 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
+#include <QMessageBox>
+#include <QSettings>
 #include "ProfileManager3.hpp"
 #include "ui_ProfileManager3.h"
+#include <Forms/Gen3/ProfileEditor3.hpp>
 
 ProfileManager3::ProfileManager3(QWidget *parent) :
-    QMainWindow(parent),
+    QWidget(parent),
     ui(new Ui::ProfileManager3)
 {
     ui->setupUi(this);
@@ -33,64 +36,38 @@ ProfileManager3::ProfileManager3(QWidget *parent) :
 
 ProfileManager3::~ProfileManager3()
 {
-    delete ui;
-    delete model;
-}
+    QSettings setting;
+    setting.setValue("profileManager3/geometry", this->saveGeometry());
 
-void ProfileManager3::changeEvent(QEvent *event)
-{
-    if (event != NULL)
-    {
-        switch (event->type())
-        {
-            case QEvent::LanguageChange:
-                ui->retranslateUi(this);
-                break;
-            default:
-                break;
-        }
-    }
+    delete ui;
 }
 
 void ProfileManager3::setupModels()
 {
-    model->setModel(Profile3::loadProfileList());
+    model = new Profile3Model(ui->tableView);
+    model->addItems(Profile3::loadProfileList());
     ui->tableView->setModel(model);
-    ui->tableView->horizontalHeader()->setSectionResizeMode(QHeaderView::Stretch);
+
+    QSettings setting;
+    if (setting.contains("profileManager3/geometry")) this->restoreGeometry(setting.value("profileManager3/geometry").toByteArray());
 }
 
 void ProfileManager3::on_pushButtonNew_clicked()
 {
-    ProfileManager3NewEdit *dialog = new ProfileManager3NewEdit();
-    connect(dialog, SIGNAL (newProfile(Profile3)), this, SLOT (registerProfile(Profile3)));
-    dialog->exec();
-}
-
-void ProfileManager3::registerProfile(Profile3 profile)
-{
-    profile.saveProfile();
-    model->addItem(profile);
-    emit updateProfiles();
-}
-
-void ProfileManager3::editProfile(Profile3 profile, Profile3 original)
-{
-    profile.updateProfile(original);
-    int r = ui->tableView->currentIndex().row();
-    model->updateProfile(profile, r);
-    emit updateProfiles();
-}
-
-void ProfileManager3::on_pushButtonOk_clicked()
-{
-    this->close();
+    QScopedPointer<ProfileEditor3> dialog(new ProfileEditor3);
+    if (dialog->exec() == QDialog::Accepted)
+    {
+        Profile3 profile = dialog->getNewProfile();
+        profile.saveProfile();
+        model->addItem(profile);
+        emit updateProfiles();
+    }
 }
 
 void ProfileManager3::on_pushButtonEdit_clicked()
 {
-    int r = ui->tableView->currentIndex().row();
-
-    if (r == -1)
+    int row = ui->tableView->currentIndex().row();
+    if (row < 0)
     {
         QMessageBox error;
         error.setText(tr("Please select a profile."));
@@ -98,16 +75,20 @@ void ProfileManager3::on_pushButtonEdit_clicked()
         return;
     }
 
-    ProfileManager3NewEdit *dialog = new ProfileManager3NewEdit(model->getProfile(r));
-    connect(dialog, SIGNAL (editProfile(Profile3, Profile3)), this, SLOT (editProfile(Profile3, Profile3)));
-    dialog->exec();
+    QScopedPointer<ProfileEditor3> dialog(new ProfileEditor3(model->getItem(row)));
+    if (dialog->exec() == QDialog::Accepted)
+    {
+        Profile3 profile = dialog->getNewProfile();
+        profile.updateProfile(dialog->getOriginal());
+        model->updateItem(profile, row);
+        emit updateProfiles();
+    }
 }
 
 void ProfileManager3::on_pushButtonDelete_clicked()
 {
-    int r = ui->tableView->currentIndex().row();
-
-    if (r == -1)
+    int row = ui->tableView->currentIndex().row();
+    if (row < 0)
     {
         QMessageBox error;
         error.setText(tr("Please select a profile."));
@@ -115,10 +96,14 @@ void ProfileManager3::on_pushButtonDelete_clicked()
         return;
     }
 
-    Profile3 profile = model->getProfile(r);
-    profile.deleteProfile();
+    QMessageBox message(QMessageBox::Question, tr("Delete profile"), tr("Are you sure you wish to delete this profile?"), QMessageBox::Yes | QMessageBox::No);
+    if (message.exec() == QMessageBox::Yes)
+    {
 
-    model->removeProfile(r);
+        Profile3 profile = model->getItem(row);
+        profile.deleteProfile();
+        model->removeItem(row);
 
-    emit updateProfiles();
+        emit updateProfiles();
+    }
 }
